Adding Custom Exceptions

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using HR.LeaveManagement.Application.Contracts.Persistence;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -13,7 +14,8 @@ public class DeleteLeaveTypeCommandHandler(ILeaveTypeRepository leaveTypeReposit
         var leaveTypeToDelete = await _leavetyperepository.GetByIdAsync(request.Id);
 
         // verify that record exists
-
+        if( leaveTypeToDelete is null)
+            throw new NotFoundException(nameof(LeaveType), request.Id);
 
         // remove from database
         await _leavetyperepository.DeleteAsync(leaveTypeToDelete);
